tests/requirements-test.txt
pytest
pytest-django
model_mommy